refactor: разбивает модель на отдельные модели
* Перемещает модели в папку global
* Передает данные в презентер хедера
