+++ src/global/model/destinations-model.js
@@ -1,0 +1,9 @@
+import { createDestinationDataset } from '../mock/destinations.js';
+
+export default class DestinationsModel {
+  #destinations = createDestinationDataset();
+
+  get destinations() {
+    return this.#destinations;
+  }
+}
