+++ src/global/model/types-model.js
@@ -1,0 +1,9 @@
+import { createEventTypeDataset } from '../mock/types.js';
+
+export default class TypesModel {
+  #types = createEventTypeDataset();
+
+  get types() {
+    return this.#types;
+  }
+}
